add useButtonStyles
